% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\docType{package}
\name{lab5}
\alias{lab5}
\alias{lab5-package}
\title{R Programming Lab5 Package}
\description{
A package which implements an R6 class costructor called "MapStamen".
   This constructor returns an object as ggmap class which is a map extracted from
   mapStamen API to be showed, saved, or subjected to any interested processing.
   The map is drived out besed on a box defined by two pairs of longitudes and
   latitudes. These coordinates are to be passed by user as the arguments.
   In addition the type of map and the required zoom can be determined and passed
   as the arguments.
}
\seealso{
[http://maps.stamen.com]
}
\author{
Ali Etminan, Mohsen Pirmoradiyan
}
