% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapStamen.R
\docType{class}
\name{MapStamen}
\alias{MapStamen}
\title{An R6 Class to download, view, and save a map based on desired coordinates.}
\format{\code{\link{R6Class}} object.}
\usage{
MapStamen(left, bottom, right, top, mapType, Zoom)
}
\arguments{
\item{left}{The lowerleft longitude of the area-box to be downloaded.}

\item{bottom}{The lowerleft latitude of the area-box to be downloaded.}

\item{right}{The upperright longitude of the area-box to be downloaded.}

\item{top}{The upperright latitude of the area-box to be downloaded.}

\item{mapType}{A string indicating the type of map("terrain", "toner",...)}

\item{Zoom}{zoom level}
}
\value{
A MapStamen class object generated by R6 class "MapStamen".
}
\description{
MapStamen is an object generator as R6 class. The returnd object is a map
   which has been downloaded from mapStamen API.
}
\details{
MapStamen generates an object as R6 class. The returnd object is a map
   which has been downloaded from mapStamen API. The object
   will be constructed by passing the lower left and upperright coordinates of the 
   area of interest. Type of map and the requiered zoom can
    also be specified as the arguments.#'
}
\section{Public Methods}{

 \describe{
   \item{\code{initializ(left, bottom, right, top =, mapType, Zoom)}}{This method will be initialized the class generator as the costructor is called. A new object will be generated.}
   \item{\code{showMap()}}{This method displays the downloaded map.}
   \item{\code{saveMap()}}{This method saves a png file in the working directory.}
   \item{\code{ClGoodday()}}{This method display the No. of healthy days in California in 2019.}
   \item{\code{ClModday()}}{This method display the No. of moderate days in California in 2019.}
   \item{\code{ClUnhealthyDay()}}{This method display the No. of unhealthy days in California in 2019.}      
     }
}

\section{Private Methods}{

 \describe{
   \item{\code{getMap()}}{This method dowloads the map from Stamen API and will be implemented in initialize method.}
   
   }
}

\examples{
myMap = MapStamen$new(left = 86.05, bottom = 27.21,
                               right = 87.81, top = 28.76, mapType = "toner-lite", Zoom = 5 )
                                   
}
\seealso{
http://maps.stamen.com
}
\keyword{datasets}
